requirements.txt
ifcopenshell
pymongo
python-dotenv